move functions to transaction utils
